pyproject.toml
[project]
name = "drosophila-agent-benchmark"
version = "0.1.0"
description = "Benchmark for evaluating AI agents on Drosophila gene annotation tasks"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2",
    "langchain-openai>=0.3",
    "langchain-community>=0.3",
    "datasets>=3.0",
    "rank-bm25>=0.2",
    "whoosh>=2.7",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    # MCP backbone agent dependencies
    "mcp>=1.0",
    "openai-agents>=0.1",
    # Evaluation dependencies
    "goatools>=1.4",
    "tqdm>=4.67.1",
    "matplotlib>=3.10.8",
    "tenacity>=9.1.2",
    "adjusttext>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[project.scripts]
run-langgraph = "agent.run_langgraph:main"
run-mcp = "agent.run_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "scripts", "eval"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["agent", "scripts"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
